+++ get_mapinfo.c
@@ -6,7 +6,7 @@
 /*   By: kishizu <kishizu@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/04 20:02:59 by kishizu           #+#    #+#             */
-/*   Updated: 2023/12/05 21:49:02 by kishizu          ###   ########.fr       */
+/*   Updated: 2023/12/06 19:52:25 by kishizu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,11 +76,9 @@ void	set_coordinates(t_info *fdf)
 		{
 			// fdf->xyz[points].initial_x = x;
 			// fdf->xyz[points].initial_y = y;
-			fdf->xyz[points].x = x;
-			fdf->xyz[points].y = y;
-			fdf->xyz[points].z = fdf->map_int[y][x];
-			// printf ("[%d][%d][%d]\n", x, y, fdf->xyz[points].z);
-			// fflush(stdout);
+			fdf->xyz[points].x = (double)x;
+			fdf->xyz[points].y = (double)y;
+			fdf->xyz[points].z = (double)fdf->map_int[y][x];
 			x++;
 			points++;
 		}
